+++ up42/tasking.py
@@ -38,6 +38,7 @@ class Tasking(catalog.CatalogBase):
     def __init__(self):
         super().__init__(glossary.CollectionType.TASKING)
 
+    @utils.deprecation("BatchOrderTemplate", "3.0.0")
     def construct_order_parameters(
         self,
         data_product_id: str,